Avoid OpenGL disabling and termination-for-restart at build-time (unit tests)
Not fully sure what is going on, but should help, as in the Windows
case LO builds and thus unit tests often are run in various remote
desktop or virtualization setups where the OpenGL support is lacking
or of an ancient version.

+++ vcl/opengl/win/gdiimpl.cxx
@@ -357,6 +357,10 @@ namespace
 
 void disableOpenGLAndTerminateForRestart()
 {
+    // Detect LO build-time
+    if (getenv("VCL_HIDE_WINDOWS"))
+        return;
+
     OpenGLZone::hardDisable();
     TerminateProcess(GetCurrentProcess(), EXITHELPER_NORMAL_RESTART);
 }
